Began moving code to abstract superclass
This has been tested and behaves correctly.

The tested version did not display the icon
correctly. I believe I have fixed the problem
with the icon but have not tested again.
